+++ src/inspecjs.ts
@@ -1,4 +1,4 @@
 // Our foreign package API.
 
 export { InspecOutput, Profile, Control } from "./types";
-export { HeimdallState } from "./store";
+export { HeimdallState } from "./heimdall_store";
